+++ packages/twenty-front/src/modules/activities/emails/components/Emails.tsx
@@ -53,9 +53,6 @@ export const Emails = ({ entity }: { entity: ActivityTargetableEntity }) => {
 
   const timelineMessages: TimelineMessage[] = messages.data.timelineMessage;
 
-  // This hard limit is just for the POC, we will implement pagination later
-  const firstTenMessages = timelineMessages.slice(0, 10);
-
   return (
     <StyledContainer>
       <Section>
@@ -69,10 +66,10 @@ export const Emails = ({ entity }: { entity: ActivityTargetableEntity }) => {
           fontColor={H1TitleFontColor.Primary}
         />
         <Card>
-          {firstTenMessages.map((message: any, index: any) => (
+          {timelineMessages.map((message: any, index: any) => (
             <EmailPreview
               key={index}
-              divider={index < firstTenMessages.length - 1}
+              divider={index < timelineMessages.length - 1}
               email={message}
             />
           ))}
